[#1114] test(library): extract tests run on GitHub to separate file
It's a preliminary step before further refactoring.
